Bruker felleskomponenter i behandlingsappen
