+++ apps/etterlatte-beregning/src/main/kotlin/Application.kt
@@ -1,5 +1,7 @@
 package no.nav.etterlatte
 
+import no.nav.etterlatte.config.ApplicationBuilder
+
 fun main() {
     val application = ApplicationBuilder()
     application.start()
