--- apps/etterlatte-beregning/src/main/kotlin/ApplicationBuilder.kt
@@ -1,120 +1,0 @@
-package no.nav.etterlatte
-
-import com.typesafe.config.Config
-import com.typesafe.config.ConfigFactory
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.okhttp.OkHttp
-import io.ktor.client.plugins.defaultRequest
-import io.ktor.client.request.header
-import io.ktor.http.ContentType
-import io.ktor.http.HttpStatusCode
-import io.ktor.serialization.jackson.JacksonConverter
-import io.ktor.server.application.Application
-import io.ktor.server.application.install
-import io.ktor.server.application.log
-import io.ktor.server.auth.Authentication
-import io.ktor.server.auth.authenticate
-import io.ktor.server.config.HoconApplicationConfig
-import io.ktor.server.plugins.callloging.CallLogging
-import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.server.plugins.statuspages.StatusPages
-import io.ktor.server.request.header
-import io.ktor.server.request.httpMethod
-import io.ktor.server.request.path
-import io.ktor.server.response.respond
-import io.ktor.server.routing.IgnoreTrailingSlash
-import io.ktor.server.routing.routing
-import no.nav.etterlatte.libs.common.logging.CORRELATION_ID
-import no.nav.etterlatte.libs.common.logging.X_CORRELATION_ID
-import no.nav.etterlatte.libs.common.logging.getCorrelationId
-import no.nav.etterlatte.libs.common.objectMapper
-import no.nav.etterlatte.model.BeregningService
-import no.nav.etterlatte.model.VilkaarsvurderingKlientImpl
-import no.nav.etterlatte.model.behandling.BehandlingKlientImpl
-import no.nav.etterlatte.model.grunnlag.GrunnlagKlientImpl
-import no.nav.helse.rapids_rivers.RapidApplication
-import no.nav.security.token.support.v2.tokenValidationSupport
-import org.slf4j.event.Level
-import java.util.*
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation as ClientContentNegotiation
-
-class ApplicationBuilder {
-    private val env = System.getenv()
-    private val properties: ApplicationProperties = ApplicationProperties.fromEnv(env)
-    private val dataSourceBuilder = DataSourceBuilder(
-        jdbcUrl = properties.jdbcUrl,
-        username = properties.dbUsername,
-        password = properties.dbPassword
-    ).apply { migrate() }
-    private val config: Config = ConfigFactory.load()
-
-    private val dataSource = dataSourceBuilder.dataSource()
-    private val beregningRepository = BeregningRepositoryImpl(dataSource)
-    private val vilkaarsvurderingKlientImpl = VilkaarsvurderingKlientImpl(config, httpClient())
-    private val grunnlagKlientImpl = GrunnlagKlientImpl(config, httpClient())
-    private val behandlingKlient = BehandlingKlientImpl(config, httpClient())
-    private val beregningService = BeregningService(
-        beregningRepository,
-        vilkaarsvurderingKlientImpl,
-        grunnlagKlientImpl,
-        behandlingKlient
-    )
-    private val rapidsConnection =
-        RapidApplication.Builder(RapidApplication.RapidApplicationConfig.fromEnv(env.withConsumerGroupId()))
-            .withKtorModule {
-                restModule(
-                    beregningService = beregningService
-                )
-            }
-            .build()
-
-    fun start() = rapidsConnection.start()
-}
-
-fun Application.restModule(
-    beregningService: BeregningService
-) {
-    install(ContentNegotiation) {
-        register(ContentType.Application.Json, JacksonConverter(objectMapper))
-    }
-
-    install(IgnoreTrailingSlash)
-
-    install(CallLogging) {
-        level = Level.INFO
-        filter { call -> !call.request.path().matches(Regex(".*/isready|.*/isalive|.*/metrics")) }
-        format { call -> "<- ${call.response.status()?.value} ${call.request.httpMethod.value} ${call.request.path()}" }
-        mdc(CORRELATION_ID) { call -> call.request.header(X_CORRELATION_ID) ?: UUID.randomUUID().toString() }
-    }
-
-    install(StatusPages) {
-        exception<Throwable> { call, cause ->
-            call.application.log.error("En feil oppstod: ${cause.message}", cause)
-            call.respond(HttpStatusCode.InternalServerError, "En intern feil har oppstått")
-        }
-    }
-
-    install(Authentication) {
-        tokenValidationSupport(config = HoconApplicationConfig(ConfigFactory.load()))
-    }
-
-    routing {
-        authenticate {
-            beregning(beregningService)
-        }
-    }
-}
-
-private fun httpClient() = HttpClient(OkHttp) {
-    install(ClientContentNegotiation) {
-        register(ContentType.Application.Json, JacksonConverter(objectMapper))
-    }
-    defaultRequest {
-        header(X_CORRELATION_ID, getCorrelationId())
-    }
-}.also { Runtime.getRuntime().addShutdownHook(Thread { it.close() }) }
-
-fun Map<String, String>.withConsumerGroupId() =
-    this.toMutableMap().apply {
-        put("KAFKA_CONSUMER_GROUP_ID", get("NAIS_APP_NAME")!!.replace("-", ""))
-    }
